Add CommonSeqBasedMessageImpl & CommonTimeBasedMessageImpl
